requirements.txt
opencv_contrib_python==4.5.1.48
streamlit==0.73.1
torch==1.7.1+cpu
torchvision==0.8.2+cpu
Pillow==8.1.2
